+++ generator/CHANGELOG.md
@@ -19,6 +19,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ---
 
+## [0.2.0] - 2025-05-01
+
+### Added
+
+- Created the initial code to generate data models and write them to disk in the approprriate NPM package.
+  Only the selected CodeSystem Enum classes are currently generated.
+- Added unit tests for all current code.
+
+### Changed
+
+- Updated project configuration and the README.md
+
+
 ## [0.1.0] - 2025-04-10
 
 ### Added